Refactor functionss to domUpdates
