+++ src/scripts.js
@@ -15,31 +15,41 @@ import './images/turing-logo.png';
 import { requestAllData, postData } from './apiCalls';
 import Destinations from './Destinations';
 import Traveler from './Traveler';
-import { renderPage, renderDestinations } from './domUpdates';
+import {
+  renderPage,
+  renderDestinations,
+  getEstimate,
+  // makeNewTrip,
+  resetForm,
+  // displayEstimate,
+  estimateContainer,
+  estimateBtn,
+  destinationSelection,
+  guests,
+  startDate,
+  daysOfTrip,
+} from './domUpdates';
 import Trip from './Trip';
 
 //global variables
 export const userID = 44;
-let currTraveler;
-let destinations;
+export let currTraveler;
+export let destinations;
 export let today = dayjs().format('YYYY/MM/DD');
-let newTrip;
+export let newTrip;
 
-const estimateBtn = document.getElementById('estimate');
-const estimateForm = document.getElementById('estimateForm');
-const estimateContainer = document.getElementById('estimateDisplay');
-const guests = document.getElementById('guests');
-const destinationSelection = document.getElementById('destinations');
-const startDate = document.getElementById('startDate');
-const daysOfTrip = document.getElementById('days');
-const confirm = document.getElementById('confirm');
-const goBack = document.getElementById('return');
+// const estimateBtn = document.getElementById('estimate');
+// const estimateForm = document.getElementById('estimateForm');
+// const estimateContainer = document.getElementById('estimateDisplay');
+// const guests = document.getElementById('guests');
+// const destinationSelection = document.getElementById('destinations');
+// const startDate = document.getElementById('startDate');
+// const daysOfTrip = document.getElementById('days');
 
 //event listener
 estimateBtn.addEventListener('click', (event) => {
   getEstimate(event);
 });
-
 estimateContainer.addEventListener('click', (event) => {
   processRequst(event);
 });
@@ -72,22 +82,22 @@ const loadPage = () => {
 
 loadPage();
 
-const getEstimate = (event) => {
-  event.preventDefault();
+// const getEstimate = (event) => {
+//   event.preventDefault();
 
-  let trip = makeNewTrip();
+//   let trip = makeNewTrip();
 
-  if (trip.travelers && dayjs(trip.date).isAfter(today) && trip.duration) {
-    estimateForm.classList.add('hidden');
-    estimateContainer.classList.remove('hidden');
+//   if (trip.travelers && dayjs(trip.date).isAfter(today) && trip.duration) {
+//     estimateForm.classList.add('hidden');
+//     estimateContainer.classList.remove('hidden');
 
-    displayEstimate(trip);
-  } else {
-    console.log('butts');
-  }
-};
+//     displayEstimate(trip);
+//   } else {
+//     console.log('butts');
+//   }
+// };
 
-const makeNewTrip = () => {
+export const makeNewTrip = () => {
   let id = destinations.list.find(
     (place) => place.destination === destinationSelection.value
   );
@@ -107,24 +117,24 @@ const makeNewTrip = () => {
   return newTrip;
 };
 
-const resetForm = () => {
-  estimateForm.classList.remove('hidden');
-  estimateContainer.classList.add('hidden');
-};
+// const resetForm = () => {
+//   estimateForm.classList.remove('hidden');
+//   estimateContainer.classList.add('hidden');
+// };
 
-const displayEstimate = (tripInfo) => {
-  estimateContainer.innerHTML = `
-    <h3>${destinationSelection.value}</h3>
-    <h3>${dayjs(tripInfo.date).format('MMMM D, YYYY')}</h3>
-    <h3>Estimated cost for ${tripInfo.travelers} travelers for ${
-    tripInfo.duration
-  } days is ${tripInfo.calculateCost(destinations)}</h3>
-    <div>
-      <button id="confirm" >Confirm Selection</button>
-      <button id="return" >Try Again</button>
-    </div>
-    `;
-};
+// // const displayEstimate = (tripInfo) => {
+// //   estimateContainer.innerHTML = `
+// //     <h3>${destinationSelection.value}</h3>
+// //     <h3>${dayjs(tripInfo.date).format('MMMM D, YYYY')}</h3>
+// //     <h3>Estimated cost for ${tripInfo.travelers} travelers for ${
+// //     tripInfo.duration
+// //   } days is ${tripInfo.calculateCost(destinations)}</h3>
+// //     <div>
+// //       <button id="confirm" >Confirm Selection</button>
+// //       <button id="return" >Try Again</button>
+// //     </div>
+// //     `;
+// // };
 
 const processRequst = (event) => {
   event.preventDefault();
